Framework #3
Update file structures and added todo
